Spliting gasometer out of interpreter
